+++ package.json
@@ -22,7 +22,7 @@
     "primeicons": "^6.0.1",
     "primeng": "^17.10.0",
     "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
+    "tslib": "^2.6.2",
     "zone.js": "~0.14.2"
   },
   "devDependencies": {
