+++ src/app/app.routes.ts
@@ -1,8 +1,10 @@
 import { Routes } from '@angular/router';
-import { AutoCadastroComponent } from './auto-cadastro/auto-cadastro.component';
+import { AutoCadastroComponent } from './components/auto-cadastro/auto-cadastro.component';
+import { LoginComponent } from './components/login/login.component';
 
 export const routes: Routes = [
-  { path: '', component: AutoCadastroComponent, pathMatch: 'full' },
+  { path: 'autocadastro', component: AutoCadastroComponent, pathMatch: 'full' },
+  { path: 'login', component: LoginComponent, pathMatch: 'full' },
 
   { path: '**', redirectTo: '', pathMatch: 'full' },
 ];
