--- src/app/auto-cadastro/auto-cadastro.component.html
@@ -1,1 +1,0 @@
-<p>auto-cadastro works!</p>
