--- src/app/auto-cadastro/auto-cadastro.component.ts
@@ -1,12 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-auto-cadastro',
-  standalone: true,
-  imports: [],
-  templateUrl: './auto-cadastro.component.html',
-  styleUrl: './auto-cadastro.component.scss'
-})
-export class AutoCadastroComponent {
-
-}
