+++ src/app/components/login/login.component.html
@@ -1,0 +1,68 @@
+<div
+  class="surface-card p-4 shadow-2 border-round w-full lg:w-6"
+  style="margin: auto"
+>
+  <div class="text-center mb-5">
+    <img
+      src="assets/images/blocks/logos/hyper.svg"
+      alt="Image"
+      height="50"
+      class="mb-3"
+    />
+    <div class="text-900 text-3xl font-medium mb-3">Bem-vindo novamente!</div>
+    <span class="text-600 font-medium line-height-3">Já possui uma conta?</span
+    ><a
+      (click)="showSignupForm = true"
+      class="font-medium no-underline ml-2 text-blue-500 cursor-pointer"
+      >Cadastre-se hoje!</a
+    >
+  </div>
+  <div>
+    <label for="email1" class="block text-900 font-medium mb-2">Email</label
+    ><input
+      id="email1"
+      type="text"
+      placeholder="Endereço de email"
+      pinputtext=""
+      class="p-inputtext p-component p-element w-full mb-3"
+    /><label for="password1" class="block text-900 font-medium mb-2"
+      >Senha</label
+    ><input
+      id="password1"
+      type="password"
+      placeholder="Senha"
+      pinputtext=""
+      class="p-inputtext p-component p-element w-full mb-3"
+    />
+    <div class="flex align-items-center justify-content-between mb-6">
+      <div class="flex align-items-center">
+        <p-checkbox id="rememberme1" styleclass="mr-2" class="p-element"
+          ><div class="mr-2 p-checkbox p-component">
+            <div class="p-hidden-accessible">
+              <input type="checkbox" value="undefined" aria-checked="false" />
+            </div>
+            <div class="p-checkbox-box"><!----></div>
+          </div>
+          <!----></p-checkbox
+        ><label for="rememberme1">Lembrar de mim</label>
+      </div>
+      <a
+        class="font-medium no-underline ml-2 text-blue-500 text-right cursor-pointer"
+        >Esqueceu sua senha?</a
+      >
+    </div>
+    <button
+      pbutton=""
+      pripple=""
+      label="Sign In"
+      icon="pi pi-user"
+      class="p-element p-ripple w-full p-button p-component"
+    >
+      <span
+        class="p-button-icon p-button-icon-left pi pi-user"
+        aria-hidden="true"
+      ></span
+      ><span class="p-button-label">Entrar</span><span class="p-ink"></span>
+    </button>
+  </div>
+</div>
