+++ src/app/components/login/login.component.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { CheckboxModule } from 'primeng/checkbox';
+import { CommonModule } from '@angular/common';
+import { DropdownModule } from 'primeng/dropdown';
+
+@Component({
+  selector: 'app-login',
+  standalone: true,
+  imports: [CheckboxModule, CommonModule, DropdownModule],
+  templateUrl: './login.component.html',
+  styleUrl: './login.component.scss',
+})
+export class LoginComponent {
+  showSignupForm = false;
+}
